break temperature/roman filter tests into separate file
